Splitting bridge_converter.
This abstracts out:
* type_converter
* typedef_analyzer

and begins to abstract out
* foreign_mod_converter.
