+++ engine/src/bridge_converter.rs
@@ -12,32 +12,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::{collections::HashMap, fmt::Display};
+use std::fmt::Display;
 
 use crate::{
     additional_cpp_generator::AdditionalNeed,
+    byvalue_checker::ByValueChecker,
     foreign_mod_converter::ForeignModConverter,
     function_wrapper::{ArgumentConversion, FunctionWrapper, FunctionWrapperPayload},
-    known_types::{is_known_type, known_type_substitute_path, should_dereference_in_cpp},
-    types::make_ident,
-};
-use crate::{
-    byvalue_checker::ByValueChecker,
-    types::Namespace,
-    unqualify::{unqualify_params, unqualify_ret_type},
-};
-use crate::{
     namespace_organizer::{NamespaceEntries, Use},
+    type_converter::TypeConverter,
+    typedef_analyzer::analyze_typedef_target,
+    typedef_analyzer::TypedefTarget,
+    types::make_ident,
+    types::Namespace,
     types::TypeName,
+    unqualify::{unqualify_params, unqualify_ret_type},
 };
 use proc_macro2::{TokenStream as TokenStream2, TokenTree};
 use quote::quote;
 use syn::punctuated::Punctuated;
-use syn::{parse::Parser, Field, GenericParam, Generics, ItemStruct};
 use syn::{
-    parse_quote, Attribute, FnArg, ForeignItem, ForeignItemFn, GenericArgument, Ident, Item,
-    ItemForeignMod, ItemMod, Pat, PathArguments, PathSegment, ReturnType, Type, TypePath, TypePtr,
-    TypeReference,
+    parse::Parser, parse_quote, Attribute, Field, FnArg, ForeignItem, ForeignItemFn, GenericParam,
+    Generics, Ident, Item, ItemForeignMod, ItemMod, ItemStruct, Pat, ReturnType, Type, TypePtr,
 };
 
 #[derive(Debug)]
@@ -134,12 +130,11 @@ impl BridgeConverter {
                     extern_c_mod_items: Vec::new(),
                     additional_cpp_needs: Vec::new(),
                     idents_found: Vec::new(),
-                    types_found: Vec::new(),
+                    type_converter: TypeConverter::new(),
                     byvalue_checker: ByValueChecker::new(),
                     pod_requests: &self.pod_requests,
                     include_list: &self.include_list,
                     final_uses: Vec::new(),
-                    typedefs: HashMap::new(),
                 };
                 conversion.convert_items(items, exclude_utilities)
             }
@@ -153,14 +148,6 @@ fn get_blank_extern_c_mod() -> ItemForeignMod {
     )
 }
 
-/// Analysis of a typedef.
-#[derive(Debug)]
-enum TypedefTarget {
-    NoArguments(TypeName),
-    HasArguments,
-    SomethingComplex,
-}
-
 /// A particular bridge conversion operation. This can really
 /// be thought of as a ton of parameters which we'd otherwise
 /// need to pass into each individual function within this file.
@@ -172,12 +159,11 @@ struct BridgeConversion<'a> {
     extern_c_mod_items: Vec<ForeignItem>,
     additional_cpp_needs: Vec<AdditionalNeed>,
     idents_found: Vec<Ident>,
-    types_found: Vec<TypeName>,
+    type_converter: TypeConverter,
     byvalue_checker: ByValueChecker,
     pod_requests: &'a Vec<TypeName>,
     include_list: &'a Vec<String>,
     final_uses: Vec<Use>,
-    typedefs: HashMap<TypeName, TypedefTarget>,
 }
 
 impl<'a> BridgeConversion<'a> {
@@ -311,9 +297,8 @@ impl<'a> BridgeConversion<'a> {
                         continue;
                     }
                     let tyname = TypeName::new(&ns, &ity.ident.to_string());
-                    let target = Self::analyze_typedef_target(ity.ty.as_ref());
+                    self.type_converter.insert_typedef(tyname, ity.ty.as_ref());
                     output_items.push(Item::Type(ity));
-                    self.typedefs.insert(tyname, target);
                 }
                 _ => return Err(ConvertError::UnexpectedItemInMod),
             }
@@ -387,20 +372,6 @@ impl<'a> BridgeConversion<'a> {
             || generics.type_params().next().is_some()
     }
 
-    fn analyze_typedef_target(ty: &Type) -> TypedefTarget {
-        match ty {
-            Type::Path(typ) => {
-                let seg = typ.path.segments.last().unwrap();
-                if seg.arguments.is_empty() {
-                    TypedefTarget::NoArguments(TypeName::from_type_path(typ))
-                } else {
-                    TypedefTarget::HasArguments
-                }
-            }
-            _ => TypedefTarget::SomethingComplex,
-        }
-    }
-
     fn find_nested_pod_types_in_mod(
         &mut self,
         items: &[Item],
@@ -418,7 +389,7 @@ impl<'a> BridgeConversion<'a> {
                         // without an actual need to do so.
                         continue;
                     }
-                    let typedef_type = Self::analyze_typedef_target(ity.ty.as_ref());
+                    let typedef_type = analyze_typedef_target(ity.ty.as_ref());
                     let name = TypeName::new(ns, &ity.ident.to_string());
                     match typedef_type {
                         TypedefTarget::NoArguments(tn) => {
@@ -522,7 +493,7 @@ impl<'a> BridgeConversion<'a> {
             }
         }));
         self.add_use(tyname.get_namespace(), &final_ident);
-        self.types_found.push(tyname);
+        self.type_converter.push(tyname);
         self.idents_found.push(final_ident);
         Ok(())
     }
@@ -924,151 +895,6 @@ impl<'a> BridgeConversion<'a> {
         Ok(result)
     }
 
-    fn convert_boxed_type(&self, ty: Box<Type>, ns: &Namespace) -> Result<Box<Type>, ConvertError> {
-        Ok(Box::new(self.convert_type(*ty, ns)?))
-    }
-
-    fn convert_type(&self, ty: Type, ns: &Namespace) -> Result<Type, ConvertError> {
-        let result = match ty {
-            Type::Path(p) => {
-                let newp = self.convert_type_path(p, ns)?;
-                // Special handling because rust_Str (as emitted by bindgen)
-                // doesn't simply get renamed to a different type _identifier_.
-                // This plain type-by-value (as far as bindgen is concerned)
-                // is actually a &str.
-                if should_dereference_in_cpp(&newp) {
-                    Type::Reference(parse_quote! {
-                        &str
-                    })
-                } else {
-                    Type::Path(newp)
-                }
-            }
-            Type::Reference(mut r) => {
-                r.elem = self.convert_boxed_type(r.elem, ns)?;
-                Type::Reference(r)
-            }
-            Type::Ptr(ptr) => Type::Reference(self.convert_ptr_to_reference(ptr, ns)?),
-            _ => ty,
-        };
-        Ok(result)
-    }
-
-    fn convert_ptr_to_reference(
-        &self,
-        ptr: TypePtr,
-        ns: &Namespace,
-    ) -> Result<TypeReference, ConvertError> {
-        let mutability = ptr.mutability;
-        let elem = self.convert_boxed_type(ptr.elem, ns)?;
-        // TODO - in the future, we should check if this is a rust::Str and throw
-        // a wobbler if not. rust::Str should only be seen _by value_ in C++
-        // headers; it manifests as &str in Rust but on the C++ side it must
-        // be a plain value. We should detect and abort.
-        Ok(parse_quote! {
-            & #mutability #elem
-        })
-    }
-
-    fn convert_type_path(
-        &self,
-        mut typ: TypePath,
-        ns: &Namespace,
-    ) -> Result<TypePath, ConvertError> {
-        if typ.path.segments.iter().next().unwrap().ident == "root" {
-            typ.path.segments = typ
-                .path
-                .segments
-                .into_iter()
-                .map(|s| -> Result<PathSegment, ConvertError> {
-                    let ident = &s.ident;
-                    let args = match s.arguments {
-                        PathArguments::AngleBracketed(mut ab) => {
-                            ab.args = self.convert_punctuated(ab.args, ns)?;
-                            PathArguments::AngleBracketed(ab)
-                        }
-                        _ => s.arguments,
-                    };
-                    Ok(parse_quote!( #ident #args ))
-                })
-                .collect::<Result<_, _>>()?;
-        } else {
-            let ty = TypeName::from_type_path(&typ);
-            // If the type looks like it is unqualified, check we know it
-            // already, and if not, qualify it according to the current
-            // namespace. This is a bit of a shortcut compared to having a full
-            // resolution pass which can search all known namespaces.
-            if !self.types_found.contains(&ty) && !is_known_type(&ty) {
-                typ.path.segments = std::iter::once(&"root".to_string())
-                    .chain(ns.iter())
-                    .map(|s| parse_quote! { #s })
-                    .chain(typ.path.segments.into_iter())
-                    .collect();
-            }
-        }
-        let mut last_seg_args = None;
-        let mut seg_iter = typ.path.segments.iter().peekable();
-        while let Some(seg) = seg_iter.next() {
-            if !seg.arguments.is_empty() {
-                if seg_iter.peek().is_some() {
-                    panic!("Did not expect bindgen to create a type with path arguments on a non-final segment")
-                } else {
-                    last_seg_args = Some(seg.arguments.clone());
-                }
-            }
-        }
-        drop(seg_iter);
-        let tn = TypeName::from_type_path(&typ);
-        // Let's see if this is a typedef.
-        let typ = self
-            .resolve_typedef(&tn)?
-            .map(|x| x.to_type_path())
-            .unwrap_or(typ);
-
-        // This will strip off any path arguments...
-        let mut typ = known_type_substitute_path(&typ).unwrap_or(typ);
-        // but then we'll put them back again as necessary.
-        if let Some(last_seg_args) = last_seg_args {
-            let last_seg = typ.path.segments.last_mut().unwrap();
-            last_seg.arguments = last_seg_args;
-        }
-        Ok(typ)
-    }
-
-    fn resolve_typedef<'b>(
-        &'b self,
-        tn: &'b TypeName,
-    ) -> Result<Option<&'b TypeName>, ConvertError> {
-        match self.typedefs.get(&tn) {
-            None => Ok(None),
-            Some(TypedefTarget::NoArguments(original_tn)) => {
-                match self.resolve_typedef(original_tn)? {
-                    None => Ok(Some(original_tn)),
-                    Some(further_resolution) => Ok(Some(further_resolution)),
-                }
-            }
-            _ => Err(ConvertError::ComplexTypedefTarget(tn.to_cpp_name())),
-        }
-    }
-
-    fn convert_punctuated<P>(
-        &self,
-        pun: Punctuated<GenericArgument, P>,
-        ns: &Namespace,
-    ) -> Result<Punctuated<GenericArgument, P>, ConvertError>
-    where
-        P: Default,
-    {
-        let mut new_pun = Punctuated::new();
-        for arg in pun.into_iter() {
-            new_pun.push(match arg {
-                GenericArgument::Type(t) => GenericArgument::Type(self.convert_type(t, ns)?),
-                _ => arg,
-            });
-        }
-        Ok(new_pun)
-    }
-
     /// Generate lots of 'use' statements to pull cxxbridge items into the output
     /// mod hierarchy according to C++ namespaces.
     fn generate_final_use_statements(&mut self) {
@@ -1096,6 +922,10 @@ impl<'a> BridgeConversion<'a> {
             );
             output_items.push(Item::Mod(new_mod));
         }
+    }
+
+    fn convert_boxed_type(&self, ty: Box<Type>, ns: &Namespace) -> Result<Box<Type>, ConvertError> {
+        self.type_converter.convert_boxed_type(ty, ns)
     }
 }
